Assets/MapLoader: do not derive from UOFileLoader
This class has really nothing to do with the UOFileLoader interface.
All it inherits is the Dispose() method which however is effectively a
no-op here because MapLoader does not override ClearResources().  To
avoid build failures, this commit adds an empty Dispose() method; code
to dispose all the contained UOFile instances should be added here.