Assets/MapLoader: check "is not UOFileUop" instead of "is UOFileMul"
As noted in the 147565dd3 commit message, MapLoader still uses
UOFileMul because ResetPatchesInBlockTable() relies on this class for
the "is" check.  This commit eliminates this kludge by inverting the
class check, making UOFileUop the special case that *disables*
ResetPatchesInBlockTable().

In fact, we only need to check wheter _filesMap[i] is (not) a
UOFileUop; the other arrays cannot ever contain a UOFileUop instance.
There, we only need the "!=null" check.