Assets/FontLoader: fix use-after-free crash bug
This fixes a bug that has been present since the very first ClassicUO
version; the faulty code was added by commit 5869b9c7898; previously,
there was some draft code without this bug, but it was removed by the
commit.

The method Load() creates 21 UOFile objects which map all 21 font
files into memory.  Then pointers into this memory mapping are stored
in _fontData and _unicodeFontAddress.  When Load() returns, the last
reference to these UOFile objects is released and they are subject to
garbage collection.

Eventually, when the garbage collector kicks in, all UOFile instances
are destructed and the memory mappings are unmapped.  Any further
access to this memory are will then either result in a crash or read
garbage, if that memory region just happens to have been allocated by
something else meanwhile.

The fix is simple: add variables holding references to these UOFile
instances.