NetClient: use ReadAsync()
Call ReadAsync() in a loop instead of calling Read() only when data is
available.  This allows detecting if the connection was closed by the
server.  Previously, this was only ever detected if Send() fails, but
if no data gets sent, we never learn about the disconnect.

For example, if the server closes the connection during AccountLogin,
the "Verifying Account" screen is stuck with no visual feedback.  This
is because the disconnect was never detected, but also because the
LoginScene does not check for disconnects.  The former is fixed by
this patch (a "Disconnected" warning gets logged), but the LoginScene
is still buggy.